+++ CHANGES.rst
@@ -16,6 +16,9 @@ Changelog
   [gbastien]
 - Highlight `Yes` value in red in `MeetingConfig.labelsConfig` datagrid field.
   [gbastien]
+- Fixed `ftw.labels` viewlet error when rendered on an item template, now item
+  templates managers can select any labels on item templates and viewlet.
+  [gbastien]
 
 4.2.28.9 (2026-01-27)
 ---------------------